Reroute bug on successful login corrected for empty returnUrl.
Add email address to activated link to prevent repeat entry.
The password reset was pulling the hash as an integer instead of a
string.
The hash is now included with the change password request.
The password patch control throws an exception if the user's hash is
wrong.
The visitor hash is reset after a successful password change.